Place coverage-ish tests for subnormal behaviour in `tests/cover/`
